% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{add_labels_column}
\alias{add_labels_column}
\title{add_labels_column}
\usage{
add_labels_column(df, labels_assocation)
}
\arguments{
\item{df}{Dataframe.}

\item{labels_assocation}{Vector of label association.}
}
\value{
Dataframe.
}
\description{
function to add label association column.
}
\examples{
\donttest{add_labels_column()}
}
