% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{assign_events_to_nearest_centroids}
\alias{assign_events_to_nearest_centroids}
\title{assign_events_to_nearest_centroids}
\usage{
assign_events_to_nearest_centroids(
  gated_df,
  n_cores = 1,
  method_dist = "euclidean",
  thr_dist = 0.15,
  include_zero = F,
  remove_centroids = T
)
}
\arguments{
\item{gated_df}{dataframe with labels (third column).}

\item{n_cores}{Number of cores. Default to 1.}

\item{method_dist}{Distance method calculation. Default to euclidean.}

\item{thr_dist}{Distance threshold for centroids calculation. Default to 0.15.}

\item{include_zero}{Consider centroid of label 0. Default to False.}

\item{remove_centroids}{Remove centroids too near each other based on thr_dist value.}
}
\value{
Dataframe.
}
\description{
function to assign events to class with nearest centroid.
}
\examples{
\donttest{assign_events_to_nearest_centroids()} 
}
