% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{check_polygons_intersection}
\alias{check_polygons_intersection}
\title{extract_polygon_gates}
\usage{
check_polygons_intersection(list_df_hull)
}
\arguments{
\item{list_df_hull}{List of polygons coordinates}
}
\value{
float
}
\description{
function to check polygons intersection.
}
\examples{
\donttest{check_polygons_intersection()}
}
