% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{compute_gates}
\alias{compute_gates}
\title{compute_gates}
\usage{
compute_gates(gated_df, list_final_polygons_coords, no_classes = F)
}
\arguments{
\item{gated_df}{dataframe with labels (third column).}

\item{list_final_polygons_coords}{List of dataframes containing polygon coordinates.}

\item{no_classes}{Generate third column of labels. Default to False.}
}
\value{
Dataframe.
}
\description{
function to assign events based on polygon gates.
}
\examples{
\donttest{compute_gates()}
}
\keyword{flowMagic}
