% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{convert_to_integers_chr}
\alias{convert_to_integers_chr}
\title{convert_to_integers_chr}
\usage{
convert_to_integers_chr(df)
}
\arguments{
\item{df}{Dataframe with classes (third column) to check and eventually modify}
}
\value{
Dataframe
}
\description{
function to text classes to integers-like classes if necessary.
}
\examples{
\donttest{convert_to_integers_chr()}
}
