% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{csv_to_dens}
\alias{csv_to_dens}
\title{csv_to_dens}
\usage{
csv_to_dens(df, with_classes = T, n_coord = "df")
}
\arguments{
\item{df}{Dataframe of marker expression values.}

\item{with_classes}{Consider classes. Default to True.}

\item{n_coord}{Grid size. Default to df.}
}
\value{
Dataframe.
}
\description{
function to get density of events (with classes associated if present).
}
\examples{
\donttest{csv_to_dens()}
}
