% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{.densRange}
\alias{.densRange}
\title{.densRange}
\usage{
.densRange(x, y, gate, pos = FALSE)
}
\arguments{
\item{x}{Coordinates of x axis.}

\item{y}{Coordinates of y axis}

\item{gate}{gate threshold.}

\item{pos}{refere to the pos in deGatePlot() function.}
}
\value{
list of numbers.
}
\description{
function to get the correct coordinates (correct range) based on the density coordinates x and y generated by the density() function.
}
\examples{
\donttest{.densRange()}
}
