% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting.R
\name{export_raw_gs_plots}
\alias{export_raw_gs_plots}
\title{export_raw_gs_plots}
\usage{
export_raw_gs_plots(
  gs,
  node_name,
  channel_x,
  channel_y,
  path_output,
  n_cores = 1,
  x_lab = NULL,
  y_lab = NULL,
  w_val = 7,
  h_val = 7,
  size_points = 1,
  return_data = FALSE,
  ...
)
}
\arguments{
\item{gs}{GatingSet}

\item{node_name}{TODOLIST}

\item{channel_x}{TODOLIST}

\item{channel_y}{TODOLIST}

\item{path_output}{Path to the directory where to export the plots.}

\item{n_cores}{Number of cores to use. Default to 1.}

\item{x_lab}{x-axis label. Default to NULL (inherited from GatingSet metadata).}

\item{y_lab}{y-axis label. Default to NULL (inherited from GatingSet metadata).}

\item{w_val}{width value. Default to 7 inches.}

\item{h_val}{height value. Default to 7 inches.}

\item{size_points}{Size points scatter plot.}

\item{return_data}{If TRUE, return the list of dataframes used to generate the plots. Default to FALSE.}

\item{...}{Additional arguments passed to the function}
}
\value{
TODOLIST
}
\description{
function to generate ungated plots from selected gs node
}
\examples{
A <- 2+2
}
