% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMagic-package.R
\docType{package}
\name{flowMagic-package}
\alias{flowMagic}
\alias{flowMagic-package}
\title{flowMagic: Automated gating of flow cytometry data}
\description{
To learn more about using the flowMagic package, please start by browsing the
vignettes:
}
\author{
\strong{Maintainer}: Sebastiano Montante \email{semont93@gmail.com}

}
\keyword{internal}
