% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{flowmagic_pred_to_poly_gates}
\alias{flowmagic_pred_to_poly_gates}
\title{flowmagic_pred_to_poly_gates}
\usage{
flowmagic_pred_to_poly_gates(
  list_df,
  pred_label,
  gate_label,
  n_cores = 1,
  concavity_val = 10
)
}
\arguments{
\item{list_df}{List of dataframes with three columns: marker 1, marker 2, class. Can get input generated by the magicPred function.}

\item{pred_label}{Select prediction label to extract polygon.}

\item{gate_label}{Label for the selected polygon within the GatingSet framework.}

\item{n_cores}{Number of cores to use. Default to 1.}

\item{concavity_val}{Concavity value for drawing polygon. Default to 10.}
}
\value{
list
}
\description{
function to label points of dataframe based on polygon coordinates
}
\examples{
\donttest{flowmagic_pred_to_poly_gates()}
}
\keyword{flowMagic}
