% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_classes_expr_df}
\alias{get_classes_expr_df}
\title{get_classes_expr_df}
\usage{
get_classes_expr_df(dens_df, original_df)
}
\arguments{
\item{dens_df}{Dataframe of density values.}

\item{original_df}{Original dataframe.}
}
\value{
Dataframe.
}
\description{
function to get classes of original expression df based on density df predictions.
}
\examples{
\donttest{get_classes_expr_df()}
}
\keyword{flowMagic}
