% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_density_features}
\alias{get_density_features}
\title{get_density_features}
\usage{
get_density_features(df_dens, min_height = 0.06)
}
\arguments{
\item{df_dens}{Dataframe of density estimates for both markers.}

\item{min_height}{Minimum height of the peaks to consider.}
}
\value{
Vector of numbers.
}
\description{
function to get density features only given a bivarite density csv.
}
\examples{
\donttest{get_density_features()}
}
\keyword{flowMagic}
