% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_density_scores}
\alias{get_density_scores}
\title{get_density_scores}
\usage{
get_density_scores(df_template, df_test, select_density_features = NULL)
}
\arguments{
\item{df_template}{Dataframe of template markers expression values.}

\item{df_test}{Dataframe of test markers expression values.}

\item{select_density_features}{Select features to use. Default to NULL.}
}
\value{
Matrix of numbers.
}
\description{
function to get scores for distance template calculation.
}
\examples{
\donttest{get_density_scores()}
}
\keyword{flowMagic}
