% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_dist_template}
\alias{get_dist_template}
\title{get_dist_template}
\usage{
get_dist_template(matrix_scores, dist_method = "euclidean")
}
\arguments{
\item{matrix_scores}{Matrix of density features generated by get_density_scores function.}

\item{dist_method}{Type of distance method calculation.}
}
\value{
Number.
}
\description{
function to get distance between template and test data.
}
\examples{
\donttest{get_dist_template()}
}
