% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_distance_loc_vs_test}
\alias{get_distance_loc_vs_test}
\title{get_distance_loc_vs_test}
\usage{
get_distance_loc_vs_test(test_df, loc_df, show_plot = "none", nboot = 50)
}
\arguments{
\item{test_df}{Dataframe with bivariate marker expression.}

\item{loc_df}{Dataframe with bivariate marker expression.}

\item{show_plot}{show density comparison. Default to none.}

\item{nboot}{Number of permutations. Default to 50.}
}
\value{
List of p-values.
}
\description{
function to compare the similiraty between current test data and local training set.
}
\examples{
\donttest{get_distance_loc_vs_test()}
}
