% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_indices_cross_val}
\alias{get_indices_cross_val}
\title{get_indices_cross_val}
\usage{
get_indices_cross_val(df_train, n_cores = 1)
}
\arguments{
\item{df_train}{Dataframe of training features generated by the get_train_data function.}

\item{n_cores}{Number of cores to use. Default to 1.}
}
\value{
List of numbers.
}
\description{
function to get indices for cross val.
}
\examples{
\donttest{get_indices_cross_val()}
}
