% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_list_df_gated_plots}
\alias{get_list_df_gated_plots}
\title{get_list_df_gated_plots}
\usage{
get_list_df_gated_plots(gs, gate_name, label_pop = NULL)
}
\arguments{
\item{gs}{GatingSet}

\item{gate_name}{Name of the Gating tree node whose gating data needs to be extracted.}

\item{label_pop}{Set label for selected gate.}
}
\value{
List.
}
\description{
function to get labelled dataframe based on selected node for each sample in gs.
}
\examples{
A <- 2+2
}
