% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_train.R
\name{get_local_train_sets}
\alias{get_local_train_sets}
\title{get_local_train_sets}
\usage{
get_local_train_sets(gh, hierarchical_tree, info_hierarchy)
}
\arguments{
\item{gh}{GatingHierarchy.}

\item{hierarchical_tree}{Dataframe of hierarchy information generated by the get_hierarchy function.}

\item{info_hierarchy}{List of hierarchy information generated by the get_hierarchy function.}
}
\value{
List of Dataframes.
}
\description{
Based on the hierarchy calculated using the get_hierarchy function, we generate the local training sets gated by the biologists.
}
\examples{
\donttest{get_local_train_sets()}
}
