% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_paths_training}
\alias{get_paths_training}
\title{get_paths_training}
\usage{
get_paths_training(df_paths, n_paths = 200, seed_n = 40)
}
\arguments{
\item{df_paths}{Dataframe of paths to training data.}

\item{n_paths}{Max number of random paths to consider for number of gates.}

\item{seed_n}{Random seed. Default to 40.}
}
\value{
Vector of characters.
}
\description{
function to get paths to train data.
}
\examples{
\donttest{get_paths_training()}
}
