% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_pops_hierarchy_list}
\alias{get_pops_hierarchy_list}
\title{get_pops_hierarchy_list}
\usage{
get_pops_hierarchy_list(hierarchical_list)
}
\arguments{
\item{hierarchical_list}{list of populations names for each level.}
}
\value{
Vector of characters.
}
\description{
function to get info from a list of hierarchical dataset (it reports all the pops for each level in a vector).
}
\examples{
\donttest{get_pops_hierarchy_list()}
}
\keyword{flowMagic}
