% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_slot_hierarchy_list}
\alias{get_slot_hierarchy_list}
\title{get_slot_hierarchy_list}
\usage{
get_slot_hierarchy_list(hierarchical_list, pop_selected)
}
\arguments{
\item{hierarchical_list}{list of populations names for each level.}

\item{pop_selected}{Population name.}
}
\value{
Element of a list.
}
\description{
function to access a slot of results from the hierarchy list based on the pop selected.
}
\examples{
\donttest{get_slot_hierarchy_list()}
}
\keyword{flowMagic}
