% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_test.R
\name{get_test_sets}
\alias{get_test_sets}
\title{get_test_sets}
\usage{
get_test_sets(fs, gh)
}
\arguments{
\item{fs}{flowSet to gate.}

\item{gh}{Gating hierarchy.}
}
\value{
List of dataframes.
}
\description{
Function to obtain the data ready to be gated (validation set or test set).
It takes as input a flowSet and generates a list of dataframes with a ML structure for each ungated fcs file.
Each dataframe is the expression matrix of the root.
}
\examples{
\donttest{get_test_sets()} 
}
\keyword{flowMagic}
