% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{get_train_data}
\alias{get_train_data}
\title{get_train_data}
\usage{
get_train_data(paths_file = NULL, df_paths = NULL, n_cores = 1,
  prop_down = NULL, remove_class = NULL)
}
\arguments{
\item{paths_file}{Vector of paths. Each path points toward a single csv file containin training info (labels and bivariate expression).}

\item{df_paths}{Dataframe containing the paths of file to read. The paths must be in the first column.}

\item{n_cores}{Number of cores. Default to 1.}

\item{prop_down}{Proportion of events to consider (dowsampling). Default to NULL.}

\item{remove_class}{Vector of classes to ignore. Default to NULL.}
}
\value{
Dataframe.
}
\description{
function to import test set in csv format.
}
\examples{
\donttest{get_train_data()}
}
