% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_weights_density_features}
\alias{get_weights_density_features}
\title{get_weights_density_features}
\usage{
get_weights_density_features(df_scores)
}
\arguments{
\item{df_scores}{Dataframe of distance scores.}
}
\value{
Vector of numbers.
}
\description{
function to calculate final score based on density features.
}
\examples{
\donttest{get_weights_density_features()}
}
