% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{import_gating_info}
\alias{import_gating_info}
\title{import_sample_gated}
\usage{
import_gating_info(path, type = "gs", group_wsp = NULL)
}
\arguments{
\item{path}{path to gs or GatingML object.}

\item{type}{type of object.}

\item{group_wsp}{Group of wsp to import.}
}
\value{
GatingSet object
}
\description{
convert a flowWorkspace or a GatingMl file of the train data folder into a gated gh (gh_train)
}
\examples{
\donttest{import_sample_gated()}
}
