% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{import_png_image}
\alias{import_png_image}
\title{import_png_image}
\usage{
import_png_image(path_img)
}
\arguments{
\item{path_img}{path to directory containing csv files to read (third column is ignored).}
}
\value{
matrix.
}
\description{
function to correctly import a png image as matrix of pixels.
}
\examples{
\donttest{import_png_image()}
}
