% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{import_reference_csv}
\alias{import_reference_csv}
\title{import_reference_csv}
\usage{
import_reference_csv(path_results, n_cores = 1)
}
\arguments{
\item{path_results}{path to directory containing the csv files  to read (with third column of labels).}

\item{n_cores}{Number of cores to use. Default to 1.}
}
\value{
list of dataframes
}
\description{
function to import plain gold standards data (no hierarchy)
}
\examples{
\donttest{import_reference_csv()}
}
