% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{import_sample_gated}
\alias{import_sample_gated}
\title{import_sample_gated}
\usage{
import_sample_gated(path, type = "gs", group_wsp = NULL,
  report_gs = F)
}
\arguments{
\item{path}{path to gs or GatingML object.}

\item{type}{type of object.}

\item{group_wsp}{Group of wsp to import.}

\item{report_gs}{report gs?}
}
\value{
GatingHierarchy object
}
\description{
convert a flowWorkspace or a GatingMl file of the train data folder into a gated gh (gh_train)
}
\examples{
\donttest{import_sample_gated()}
}
