% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{import_test_set}
\alias{import_test_set}
\title{import_test_set}
\usage{
import_test_set(path, n_samples = "All", ref_f_n = 1)
}
\arguments{
\item{path}{path of directory containig the fcs files.}

\item{n_samples}{Number of samples. Default to All.}

\item{ref_f_n}{Set reference flowFrame to match channel names. Default to 1(first flowFrame).}
}
\value{
flowSet.
}
\description{
read the ungated fcs files into a flowSet.
The ungated fcs are assumed to be already cleaned,compensated,and transformed.
}
\examples{
\donttest{import_test_set()}
}
