% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{import_test_set_csv}
\alias{import_test_set_csv}
\title{import_test_set_csv}
\usage{
import_test_set_csv(path_data, n_cores = 1, xy_col = T)
}
\arguments{
\item{path_data}{path to directory containing csv files to read (third column is ignored).}

\item{n_cores}{Number cores. Default to 1.}

\item{xy_col}{Colnames equal to x and y. Default to True.}
}
\value{
List of dataframes.
}
\description{
function to import test set in csv format.
}
\examples{
\donttest{import_test_set_csv()}
}
\keyword{flowMagic}
