% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{magic_label_poly}
\alias{magic_label_poly}
\title{magic_label_poly}
\usage{
magic_label_poly(df, polygon_df, label_pol = "1")
}
\arguments{
\item{df}{Dataframe composed of two columns for marker expression of first (x axis = first column) and second marker (y axis = second column)}

\item{polygon_df}{Dataframe containing the x coordinates (first column) and y coordinates (second column) of the current polygon to label.}

\item{label_pol}{Label for current polygon.}
}
\value{
Dataframe
}
\description{
function to label points of dataframe based on polygon coordinates
}
\examples{
\donttest{magic_label_poly()}
}
\keyword{flowMagic}
