% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{magic_label_rectangle}
\alias{magic_label_rectangle}
\title{magic_label_rectangle}
\usage{
magic_label_rectangle(df, x_min, x_max, y_min, y_max, label_pol = "1")
}
\arguments{
\item{df}{Dataframe composed of two columns for marker expression of first (x axis = first column) and second marker (y axis = second column)}

\item{x_min}{x coordinate minimum}

\item{x_max}{x coordinate maximum}

\item{y_min}{y coordinate minimum}

\item{y_max}{y coordinate maximum}

\item{label_pol}{Label for current polygon.}
}
\value{
Dataframe
}
\description{
function to label points of dataframe based on rectangle coordinates
}
\examples{
\donttest{magic_label_rectangle()}
}
