% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{magic_plot_wrap}
\alias{magic_plot_wrap}
\title{magic_plot_wrap}
\usage{
magic_plot_wrap(list_gated_data, n_col_wrap = 3, size_title = 10, ...)
}
\arguments{
\item{list_gated_data}{List of dataframes composed by three columns (could be generated by the get_list_df_gated_plots() function)}

\item{n_col_wrap}{number of columns in the wrapped plot. Default to 3.}

\item{size_title}{size of title of each plot. Default to 10.}
}
\value{
wrapped plot
}
\description{
function to wrap all plots in one plot
}
\examples{
\donttest{magic_plot_wrap()}
}
