% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{magicGating}
\alias{magicGating}
\title{magicGating}
\usage{
magicGating(
  fs,
  sample_id = 1,
  channel_x,
  channel_y,
  gs_node = NULL,
  label_pol = "1",
  ...
)
}
\arguments{
\item{fs}{An object of class flowSet. Can be also an object of class cytoset (it will be converted to flowSet) or a GatingSet object (in this case, gs_node is mandatory)}

\item{sample_id}{Names of the samples to gate. It can also be the numerical index of the sample. Default to sample 1.}

\item{channel_x}{Name of the channel (x-axis).}

\item{channel_y}{Name of the channel (y-axis).}

\item{gs_node}{Name of the node to extract data, if fs is a GatingSet object this a mandatory argument.}

\item{label_pol}{Label of the gate polygon. Default to "1".}
}
\value{
List of two objects of class List.
list_poly_gates: List of polygon coordinates. Each element refers to the coordinates of one sample.
list_gated_data: List of gated data. Each element refers to the gated data of one sample.
}
\description{
function to manually gate samples in a flowSet.
}
\examples{
\donttest{magicGating()}
}
\keyword{flowMagic}
