% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{magicplot_3D}
\alias{magicplot_3D}
\title{magicplot_3D}
\usage{
magicplot_3D(
  df,
  class_col = F,
  x_lab = "x",
  y_lab = "y",
  z_lab = "z",
  type = "ML",
  size_p = 1
)
}
\arguments{
\item{df}{input dataframe composed of at least three columns. Four columns if gate needs to be plot.}

\item{class_col}{Gate to plot? Default to False.}

\item{x_lab}{label of the x axis}

\item{y_lab}{Label of the y axis}

\item{z_lab}{Label of the z axis}

\item{type}{Assuming class_col==T, If type==ML, Generate a plot colored based on the gate assignment. If type=="mesh", generate a 3d polygon gate.}

\item{size_p}{size of scatter plot points. Default to 1.}
}
\value{
plotly plot
}
\description{
function to make a 3D plot of data.
}
\examples{
\donttest{magicplot_3D()}
}
\keyword{flowMagic}
