% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{magicPlot_fs}
\alias{magicPlot_fs}
\title{magicPlot_fs}
\usage{
magicPlot_fs(fs, sample_id, channel_x, channel_y, ...)
}
\arguments{
\item{fs}{object of class flowSet or flowFrame.}

\item{sample_id}{Name of the sample data to plot. No effect if fs is an object of class flowFrame.}

\item{channel_x}{Name of the channel to plot (x-axis).}

\item{channel_y}{Name of the channel to plot (y-axis).}
}
\value{
Object of class ggplot
}
\description{
function to plot FCM data directly from objects of class flowSet or flowFrame. The data is shown in a density scatter plot.
}
\examples{
\donttest{magicPlot_fs()}
}
\keyword{flowMagic}
