% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{magicPlot_template}
\alias{magicPlot_template}
\title{magicPlot_template}
\usage{
magicPlot_template(df, size_points = 1)
}
\arguments{
\item{df}{input dataframe composed of two columns reporting markers expression (first column = marker 1, second column = marker 2).}

\item{size_points}{Size of points in scatter plot. Default to 1.}
}
\value{
Dataframe
}
\description{
function to interactively generate a template based on bivariate ungated data.
}
\examples{
\donttest{magicPlot_template()}
}
