% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{magicPred}
\alias{magicPred}
\title{magicPred}
\usage{
magicPred(test_data, magic_model, ref_model_info = NULL, n_cores = 1,
  ref_data_train = NULL, prop_down = 1)
}
\arguments{
\item{test_data}{Dataframe of test data to gate. It has only the two columns of marker expression.}

\item{magic_model}{Global trained model.}

\item{ref_model_info}{Template model}

\item{n_cores}{Number of cores to use. Default to 1.}

\item{ref_data_train}{Template data.}

\item{prop_down}{Proportion for downsampling. Default to 1 (no downsampling).}
}
\value{
List of Dataframes.
}
\description{
function to predict on plain test data (no hierarchy)
}
\examples{
\donttest{magicPred()}
}
