% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{magicTrain_dt}
\alias{magicTrain_dt}
\title{magicTrain_dt}
\usage{
magicTrain_dt(Xtrain, Ytrain, k_cv = 10, list_index = NULL,
  tune_lenght = 10)
}
\arguments{
\item{Xtrain}{Dataframe of training features.}

\item{Ytrain}{Dataframe of labels (one column).}

\item{k_cv}{Number of k for cross-validation (if method control=cv)}

\item{list_index}{List of vector of indices to use in training.}

\item{tune_lenght}{Number of parameters values trained during cross-validation.}
}
\value{
model object.
}
\description{
function to generate a decision tree training model.
}
\examples{
\donttest{magicTrain_dt()}
}
