% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{magicTrain_hierarchy}
\alias{magicTrain_hierarchy}
\title{magicTrain_local}
\usage{
magicTrain_hierarchy(
  list_train_sets,
  n_tree = 10,
  train_model = "rf",
  method_control = "oob",
  n_cores = 1
)
}
\arguments{
\item{list_train_sets}{List of labeled dataframe to train generated by the get_local_train function.}

\item{n_tree}{Number of tree for random forest model.}

\item{train_model}{Type of training model. Default to random forest ("rf").}

\item{method_control}{Cross-validation method. Default to out-of-the-bag method (oob).}

\item{n_cores}{Number of cores to use. Default to 1.}
}
\value{
List of models objects.
}
\description{
function to generate the local training models using the list of hierarchical training set.
}
\examples{
\donttest{magicTrain_local()}
}
