% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{magicTrain_nb}
\alias{magicTrain_nb}
\title{magicTrain_nb}
\usage{
magicTrain_nb(Xtrain, Ytrain, k_cv = 10)
}
\arguments{
\item{Xtrain}{Dataframe of training features.}

\item{Ytrain}{Dataframe of labels (one column).}

\item{k_cv}{Number of k for cross-validation (if method control=cv).}
}
\value{
model object.
}
\description{
function to generate a naive bayes training model.
}
\examples{
\donttest{magicTrain_nb()}
}
