% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_train.R
\name{map_to_parent}
\alias{map_to_parent}
\title{map_to_parent}
\usage{
map_to_parent(gh, binary_df)
}
\arguments{
\item{gh}{GatingHierarchy.}

\item{binary_df}{Dataframe generated by the map_to_root function.}
}
\value{
Dataframe.
}
\description{
function to get binary dataset with gate assignation based on the mother population of the selected pop (instead of the root).
}
\examples{
\donttest{map_to_parent()}
}
