% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_train.R
\name{map_to_root}
\alias{map_to_root}
\title{map_to_root}
\usage{
map_to_root(gh, pop)
}
\arguments{
\item{gh}{GatingHierarchy.}

\item{pop}{Name of the population to get events assignments.}
}
\value{
Dataframe.
}
\description{
it generates a dataset indicating what cells belong to the selected pop (1) 0 otherwise.
Old name: pre_process_manual_binary()
The dataset is always the dataset of the Root pop.
gh and pop are mandatory arguments, dim and mode have a default value.
}
\examples{
A <- 2+2
}
