% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_train.R
\name{name_pop_gating}
\alias{name_pop_gating}
\title{name_pop_gating}
\usage{
name_pop_gating(gh)
}
\arguments{
\item{gh}{GatingHierarchy.}
}
\value{
Vector of characters.
}
\description{
function to get the name of all the pops of the gating hierarchy.
}
\examples{
\donttest{name_pop_gating()}
}
\keyword{flowMagic}
