% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{post_process_gates}
\alias{post_process_gates}
\title{post_process_gates}
\usage{
post_process_gates(
  gated_df,
  n_cores = 1,
  thr_dist = 0.15,
  include_zero = F,
  remove_centroids = T,
  type = "dist",
  concavity_val = 5,
  normalize_data = T
)
}
\arguments{
\item{gated_df}{dataframe with labels (third column).}

\item{n_cores}{Number of cores. Default to 1.}

\item{thr_dist}{Distance threshold for centroids calculation. Default to 0.15.}

\item{include_zero}{Consider centroid of label 0. Default to False.}

\item{remove_centroids}{Remove centroids too near each other based on thr_dist value.}

\item{type}{Type of post-processing.}

\item{concavity_val}{Concavity of polygons for the "polygon" type of post-processing}
}
\value{
Dataframe.
}
\description{
function to post process the events after model prediction.
}
\examples{
\donttest{post_process_gates()}
}
