% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_train.R
\name{pre_process_manual_binary}
\alias{pre_process_manual_binary}
\title{pre_process_manual_binary}
\usage{
pre_process_manual_binary(gh, pop)
}
\arguments{
\item{gh}{GatingHierarchy.}

\item{pop}{Name of the population to get events assignments.}
}
\value{
Dataframe.
}
\description{
In other words it generates the dataset that indicates what cells belong to the selected pop (1) 0 otherwise.
The dataset is always the dataset of the Root pop.
gh and pop are mandatory arguments, dim and mode have a default value.
}
\examples{
\donttest{pre_process_manual_binary()}
}
