% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{process_test_data}
\alias{process_test_data}
\title{process_test_data}
\usage{
process_test_data(test_data, prop_down = 1)
}
\arguments{
\item{test_data}{Dataframe of bivariate markers expression.}

\item{prop_down}{Proportion of events (downsampling). Default to 1 (no downsampling).}
}
\value{
Dataframe.
}
\description{
function to get test data correctly formatted.
}
\examples{
\donttest{process_test_data()}
}
