% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{range01}
\alias{range01}
\title{range01}
\usage{
range01(x)
}
\arguments{
\item{x}{Vector of numbers to scale.}
}
\value{
Vector of numbers.
}
\description{
function to put data in range 0-1.
}
\examples{
\donttest{range01()}
}
\keyword{flowMagic}
