% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{smooth_hull}
\alias{smooth_hull}
\title{smooth_hull}
\usage{
smooth_hull(hull_df, spar = 0.7)
}
\arguments{
\item{hull_df}{Dataframe generate by  concaveman functions inside get_huget_hull_all_gates function}

\item{spar}{Spar value to regulate smoothing process: higher value (max 1) higher smoothing.}
}
\value{
Dataframe.
}
\description{
function to smooth concave polygons
}
\examples{
\donttest{smooth_hull()} 
}
