% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_train.R
\name{subsetting_binary_df}
\alias{subsetting_binary_df}
\title{subsetting_binary_df}
\usage{
subsetting_binary_df(gh, binary_df)
}
\arguments{
\item{gh}{GatingHierarchy.}

\item{binary_df}{Dataframe generated by the pre_process_manual_binary function.}
}
\value{
Dataframe.
}
\description{
function to get binary dataset with gate assignation based on the mother population of the selected pop (instead of the root).
}
\examples{
\donttest{subsetting_binary_df()}
}
