% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{update_label_association}
\alias{update_label_association}
\title{update_label_association}
\usage{
update_label_association(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
Vector.
}
\description{
function to update label association.
}
\examples{
\donttest{update_label_association()}
}
