% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting.R
\name{export_raw_gs_plots}
\alias{export_raw_gs_plots}
\title{exports_plots}
\usage{
export_raw_gs_plots(
  gs,
  node_name,
  channel_x,
  channel_y,
  path_output,
  n_cores = 1,
  x_lab = "x",
  y_lab = "y",
  w_val = 7,
  h_val = 7,
  size_points = 1,
  ...
)
}
\arguments{
\item{gs}{GatingSet}

\item{path_output}{Path to the directory where to export the plots.}

\item{n_cores}{Number of cores to use. Default to 1.}

\item{x_lab}{x-axis label.}

\item{y_lab}{y-axis label.}

\item{w_val}{width value. Default to 7 inches.}

\item{h_val}{height value. Default to 7 inches.}

\item{size_points}{Size points scatter plot.}
}
\description{
function to generate ungated plots from selected gs node
}
\examples{
\donttest{export_raw_gs_plots()}
}
