% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting.R
\name{exports_plots}
\alias{exports_plots}
\title{exports_plots}
\usage{
exports_plots(list_gated_data, path_output, n_cores = 1,
  type_plot = "dens", show_legend = T, x_lab = "x", y_lab = "y",
  size_title_x = 23, size_title_y = 23)
}
\arguments{
\item{list_gated_data}{list of dataframes. Each dataframe has 3 columns: marker 1 values, marker 2 values and label column.}

\item{path_output}{Path to the directory where to export the plots.}

\item{n_cores}{Path to directory containing the expression data of the files analyzed.}

\item{type_plot}{the user can choose between density (="dens) or label assignment visualization (="ML")}

\item{show_legend}{If True it shows the legend for the label assignment visualization. Default to True.}

\item{x_lab}{x-axis label.}

\item{y_lab}{y-axis label.}

\item{size_title_x}{Size x axis label.}

\item{size_title_y}{Size y axis label.}
}
\description{
function to generate plots (no hierarchy).
}
\examples{
\donttest{exports_plots()}
}
