% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting.R
\name{exports_plots}
\alias{exports_plots}
\title{exports_plots}
\usage{
exports_plots(
  list_gated_data,
  path_output,
  n_cores = 1,
  type_plot = "dens",
  show_legend = T,
  x_lab = "x",
  y_lab = "y",
  size_title_x = 23,
  size_title_y = 23,
  aspect_ratio = NULL,
  w_val = 7,
  h_val = 7,
  size_axis_text = 25,
  export_csv = F,
  ...
)
}
\arguments{
\item{list_gated_data}{list of dataframes. Each dataframe has 3 columns: marker 1 values, marker 2 values and label column.}

\item{path_output}{Path to the directory where to export the plots.}

\item{n_cores}{Number of cores to use. Default to 1.}

\item{type_plot}{the user can choose between density (="dens) or label assignment visualization (="ML")}

\item{show_legend}{If True it shows the legend for the label assignment visualization. Default to True.}

\item{x_lab}{x-axis label.}

\item{y_lab}{y-axis label.}

\item{size_title_x}{Size x axis label.}

\item{size_title_y}{Size y axis label.}

\item{aspect_ratio}{Set aspect ratio. Default to NULL> If = 1, y and x axis ticks have same distance.}

\item{w_val}{width value. Default to 7.}

\item{h_val}{height value. Default to 7.}

\item{size_axis_text}{Size of ticks labels.}

\item{export_csv}{Export plot data as csv files. Default to False.}
}
\description{
function to generate plots (no hierarchy) from list of labelled dataframes.
}
\examples{
\donttest{exports_plots()}
}
