% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{extract_polygon_gates}
\alias{extract_polygon_gates}
\title{extract_polygon_gates}
\usage{
extract_polygon_gates(gated_df, concavity_val = 1, ...)
}
\arguments{
\item{gated_df}{dataframe with labels (third column).}

\item{concavity_val}{Concavity of polygons. Default to 1.}
}
\value{
List of dataframes.
}
\description{
function to extract the polygon gates objects based on the convex hull and classes.
}
\examples{
\donttest{extract_polygon_gates()}
}
\keyword{flowMagic}
