% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{flowmagic_pred_to_gs}
\alias{flowmagic_pred_to_gs}
\title{flowmagic_pred_to_gs}
\usage{
flowmagic_pred_to_gs(list_poly_gates, gs, parent_node)
}
\arguments{
\item{list_poly_gates}{List of polygonGate objects generated using the flowmagic_pred_to_poly_gates}

\item{gs}{GatingSet to update with the flowMagic polygons.}

\item{parent_node}{Parent population name of the new gate.}
}
\value{
GatingSet
}
\description{
function to label points of dataframe based on polygon coordinates
}
\examples{
\donttest{flowmagic_pred_to_gs()}
}
