% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_flowframe_from_gs}
\alias{get_flowframe_from_gs}
\title{get_flowframe_from_gs}
\usage{
get_flowframe_from_gs(gs, node_name, sample_id)
}
\arguments{
\item{gs}{GatingSet}

\item{node_name}{Name of the Gating tree node whose gating data needs to be extracted.}

\item{sample_id}{Name or index of the sample to extract.}
}
\value{
List.
}
\description{
function to get flowframe from gs (converting cytoframe to flowframe)
}
\examples{
\donttest{get_flowframe_from_gs()}
}
\keyword{flowMagic}
