% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_train.R
\name{get_hierarchy_all_pops}
\alias{get_hierarchy_all_pops}
\title{get_hierarchy_all_pops}
\usage{
get_hierarchy_all_pops(
  gh,
  export_visnet = FALSE,
  path.output = "None",
  seed_n = 123
)
}
\arguments{
\item{gh}{GatingHierarchy.}

\item{export_visnet}{If true, it export visnetwork object.}

\item{path.output}{Path to save visnetwork object. Default to None.}

\item{seed_n}{Numeric for set.seed, default is 123}
}
\value{
List of Dataframes.
}
\description{
function to get the hierarchy of all pops from the sample manually gated (the input gating hierarchy).
Based on its output the function magicTrain will perform the training step using 
the sample manually gated (local training set) and the project discovery data (global training set).
}
\examples{
A <- 2+2
}
