% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_hierarchy_all_pops}
\alias{get_hierarchy_all_pops}
\title{get_hierarchy_all_pops}
\usage{
get_hierarchy_all_pops(gh, export_visnet = F, path.output = "None")
}
\arguments{
\item{gh}{GatingHierarchy.}

\item{export_visnet}{If true, it export visnetwork object.}

\item{path.output}{Path to save visnetwork object. Default to None.}
}
\value{
List of Dataframes.
}
\description{
function to plot hierarchy of all nodes/cell populations from a GatingHierarchy object.
}
\examples{
\donttest{get_hierarchy_all_pops()}
}
\keyword{flowMagic}
