% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{get_hull_all_gates}
\alias{get_hull_all_gates}
\title{get_hull_all_gates}
\usage{
get_hull_all_gates(gated_df, concavity_val = 1, spar_val = 0.7, smoothing = F)
}
\arguments{
\item{gated_df}{dataframe with labels (third column).}

\item{concavity_val}{Values of concavity. Default to 1.}

\item{spar_val}{Value of spar. Default to 0.7.}

\item{smoothing}{Apply smoothing of lines? Default to F.}
}
\value{
List of dataframes.
}
\description{
function to get the convex hull of all gates.
}
\examples{
\donttest{get_hull_all_gates()}
}
