% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_indices_cross_val}
\alias{get_indices_cross_val}
\title{get_indices_cross_val}
\usage{
get_indices_cross_val(df_train, n_cores = 1, train_inds = "plot_num",
  val_inds = "none", n_train_plots = 5, n_folds = 5, seed = 40,
  n_val_plots = 5)
}
\arguments{
\item{df_train}{Dataframe of training features generated by the get_train_data function.}

\item{n_cores}{Number of cores to use. Default to 1.}

\item{train_inds}{Type of method to extract training indices:  plot_num,rand_set_num,rand_set_n_gates_info.}

\item{val_inds}{Type of method to extract validation indices:  plot_num,rand_set_num,rand_set_n_gates_info.}

\item{n_train_plots}{Number of training plots to include in training data for each iteration.}

\item{n_folds}{Number of training iterations.}

\item{seed}{Seed to randomly extract data.}

\item{n_val_plots}{Number of training plots to include in validation data for each iteration.}
}
\value{
List of integers.
}
\description{
function to get indices for cross val.
}
\examples{
\donttest{get_indices_cross_val()}
}
