% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_train.R
\name{get_pop_multiclass}
\alias{get_pop_multiclass}
\title{get_pop_multiclass}
\usage{
get_pop_multiclass(gh, pop)
}
\arguments{
\item{gh}{GatingHierarchy.}

\item{pop}{Name of population to get events assignments.}
}
\value{
Vector of characters.
}
\description{
function to get the children of the selected population,
useful for multiclass classification. The output is only the pops with same dimensions
}
\examples{
\donttest{get_pop_multiclass()}
}
\keyword{flowMagic}
