% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{magicPlot}
\alias{magicPlot}
\title{magicPlot}
\usage{
magicPlot(
  df,
  type = "dens",
  polygons_coords_list = NULL,
  show_legend = T,
  size_axis_text = 18,
  size_title_x = 20,
  size_title_y = 20,
  treat_0_as_gate = F,
  x_lab = "x",
  y_lab = "y",
  gates_to_plot = NULL,
  apply_manual_scale = F,
  hull_only = F,
  size_points = 1,
  concavity_val = 20,
  aspect_ratio = NULL,
  x_lim1 = NULL,
  x_lim2 = NULL,
  y_lim1 = NULL,
  y_lim2 = NULL,
  add_labels = F,
  map_label_polygon = NULL,
  size_pol_name = 6,
  show_marginals = F,
  ...
)
}
\arguments{
\item{df}{Dataframe of bivariate markers expression (with labels if gates to plot).}

\item{type}{Type of plot to generated. "dens"=bivariate density plot. "ML"=events assignments plot}

\item{polygons_coords_list}{list of gates coordinates. Needed if labels not included in df. Default to NULL.}

\item{show_legend}{Show legend if type="ML". Default to True.}

\item{size_axis_text}{Size of axis ticks labels. Default to 18.}

\item{size_title_x}{Size of x axis label title.}

\item{size_title_y}{Size of y axis label title.}

\item{treat_0_as_gate}{Treat 0 label as gate. Defaul to False (0 label is background)}

\item{x_lab}{Label of x axis.}

\item{y_lab}{Label off y axis.}

\item{gates_to_plot}{Select labels to plot.}

\item{apply_manual_scale}{Apply predifined scale of colors. Default to True.}

\item{size_points}{Size of points in scatter plot.}

\item{concavity_val}{Concavity value. Default to 5. Higher value, less jagged boundaries}

\item{aspect_ratio}{Aspect ratio value. If = 1, y and x axis have equal distance between the ticks labels. Default to NULL.}

\item{x_lim1}{Minimum limit x axis. Default to NULL.}

\item{x_lim2}{Max limit x axis. Default to NULL.}

\item{y_lim1}{Minimum limit y axis. Default to NULL.}

\item{y_lim2}{Max limit y axis. Default to NULL.}

\item{add_labels}{add polygon labels. Default to FALSE.}

\item{map_label_polygon}{map of polygon labels (to assign custom labels). Default to NULL.}

\item{size_pol_name}{size polygon labels. Default to 6.}

\item{show_marginals}{show 1D density next to axis. Default to False.}
}
\value{
ggplot.
}
\description{
function to generate the scatter plot with colored density of the events.
}
\examples{
\donttest{magicPlot()}
}
