% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{magicPred}
\alias{magicPred}
\title{magicPred}
\usage{
magicPred(test_data, magic_model = NULL, magic_model_n_gates = NULL,
  ref_model_info = NULL, n_cores = 1, ref_data_train = NULL,
  prop_down = NULL, thr_dist = 0.05, n_points_per_plot = NULL,
  normalize_data = T, include_zero_val = T)
}
\arguments{
\item{test_data}{Dataframe of test data to gate. It has only the two columns of marker expression.}

\item{magic_model}{Global trained model to predict gates. It can be a single model or list of named models (each model trained on selected number of gates).}

\item{magic_model_n_gates}{Global trained model to predict number of gates. If different from NULL, magic_model is expected to be a list of models to predict certain gates (e.g., 5 models for 2,3,4,5 or 6 gates).}

\item{ref_model_info}{Template model to predic gates.}

\item{n_cores}{Number of cores to use. Default to 1.}

\item{ref_data_train}{Template data used to generate ref_model_info. Needed to calculate target-template distance.}

\item{prop_down}{Proportion for downsampling. Default to NULL (automatic downsampling using n_points_per_plot).}

\item{n_points_per_plot}{Number of points to consider for downsampling. Default to 500.}

\item{normalize_data}{If True, data is normalized to 0-1 range. Default to True.}

\item{include_zero_val}{considering events labeled as 0 as an additional gate when there is only one gate. Default to  True.}
}
\value{
List of Dataframes.
}
\description{
function to predict on plain test data (no hierarchy)
}
\examples{
\donttest{magicPred()}
}
