% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{magicPred_all}
\alias{magicPred_all}
\title{magicPred_all}
\usage{
magicPred_all(list_test_data, magic_model, ref_model_info = NULL,
  n_cores = 1, ref_data_train = NULL, verbose = F)
}
\arguments{
\item{list_test_data}{List of unlabeled  dataframes. It has only the two columns of marker expression.}

\item{magic_model}{Global trained model.}

\item{ref_model_info}{Template model.}

\item{n_cores}{Number of cores to use. Default to 1.}

\item{ref_data_train}{Template data.}

\item{verbose}{If True, show messages. Default to False.}
}
\value{
List of Dataframes.
}
\description{
function to predict on plain test data (no hierarchy)
}
\examples{
\donttest{magicPred_all()}
}
