% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{magicPred_all}
\alias{magicPred_all}
\title{magicPred_all}
\usage{
magicPred_all(list_test_data, magic_model = NULL,
  ref_model_info = NULL, magic_model_n_gates = NULL,
  ref_data_train = NULL, prop_down = NULL, n_points_per_plot = NULL,
  thr_dist = 0.05, n_cores = 1, normalize_data = T,
  include_zero_val = T, n_cores_all = 1)
}
\arguments{
\item{list_test_data}{List of unlabeled  dataframes. It has only the two columns of marker expression.}

\item{magic_model}{Global trained model to predict gates. It can be a single model or list of named models (each model trained on selected number of gates).}

\item{ref_model_info}{Template model to predic gates.}

\item{magic_model_n_gates}{Global trained model to predict number of gates. If different from NULL, magic_model is expected to be a list of models to predict certain gates (e.g., 5 models for 2,3,4,5 or 6 gates).}

\item{ref_data_train}{Template data used to generate ref_model_info. Needed to calculate target-template distance.}

\item{n_points_per_plot}{Number of points to consider for downsampling. Default to 500.}

\item{n_cores}{Number of cores to use to process one sample. Default to 1.}

\item{normalize_data}{If True, data is normalized to 0-1 range. Default to True.}

\item{include_zero_val}{considering events labeled as 0 as an additional gate when there is only one gate. Default to  True.}

\item{n_cores_all}{Number of cores to use across all samples. Default to 1.}
}
\value{
List of Dataframes.
}
\description{
function to predict on plain test data (no hierarchy)
}
\examples{
\donttest{magicPred_all()}
}
