% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{magicTrain_nnet}
\alias{magicTrain_nnet}
\title{magicTrain_nnet}
\usage{
magicTrain_nnet(
  Xtrain,
  Ytrain,
  k_cv = 10,
  list_index_train = NULL,
  list_index_val = NULL,
  size = 100,
  decay = 0.1,
  tune_lenght = 5
)
}
\arguments{
\item{Xtrain}{Dataframe of training features.}

\item{Ytrain}{Dataframe of labels (one column).}

\item{k_cv}{Number of k for cross-validation.}

\item{list_index_train}{List of vector of indices to use in training for each fold.}

\item{list_index_val}{List of vector of indices to use as held out data for each fold.}

\item{size}{Number of units in hidden layer (if train_model=nnet).}

\item{decay}{Decay parameter value for nnet model.}
}
\value{
model object.
}
\description{
function to generate a neural net training model.
}
\examples{
\donttest{magicTrain_nnet()}
}
