% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{merge_magicGating_labels}
\alias{merge_magicGating_labels}
\title{merge_magicGating_labels}
\usage{
merge_magicGating_labels(list_out_1, list_out_2)
}
\arguments{
\item{list_out_1}{List of Dataframes to update.}

\item{list_out_2}{List of Dataframes to merge with list_out_1 dataframes.s}
}
\value{
list of Dataframes
}
\description{
function to merge magicGating outputs when gating multiple gates in templates.
}
\examples{
\donttest{merge_magicGating_labels()}
}
\keyword{flowMagic}
