% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{process_test_data}
\alias{process_test_data}
\title{process_test_data}
\usage{
process_test_data(test_data, prop_down = NULL, n_points_per_plot = 500,
  normalize_data = T)
}
\arguments{
\item{test_data}{Dataframe of bivariate markers expression.}

\item{prop_down}{Proportion of events (downsampling). Default to NULL (downsampling using number of points).}

\item{n_points_per_plot}{Number of points for downsampling.}

\item{normalize_data}{If True, data is normalized to 0-1 range. Default to True.}
}
\value{
Dataframe.
}
\description{
function to get test data correctly formatted.
}
\examples{
\donttest{process_test_data()}
}
